* device/lib/_divulong.c: cosmetic, use y0 * device/lib/_fsadd.c: cosmetic * device/lib/clean.mk: clean pic14 and tlcs90 * device/lib/ds390/tinibios.c (_sdcc_external_startup): do not change stack * device/lib/Makefile.in: use TAB's * device/non-free/lib/clean.mk: clean pic14 * device/non-free/lib/pic14/Makefile.in, * sdas/as6808/clean.mk, * sdas/as8051/clean.mk, * sdas/as8xcxxx/clean.mk, * sdas/asgb/clean.mk, * sdas/asrab/clean.mk, * sdas/asstm8/clean.mk, * sdas/astlcs90/clean.mk, * sdas/asz80/clean.mk, * support/sdbinutils/Makefile.in: prevent warnings * sim/ucsim/cmd.src/clean.mk: clean cmdpars.* & cmdlex.* * src/clean.mk: clean pic14 and stm8 * src/common.h: cosmetic * src/ds390/main.c (_ds390_genAssemblerPreamble): added dpl, dph, acon,   (_ds390_genIVT): change addressing mode and stack at __reset_vect,   (_ds390_genInitStartup): no need for 10bit stack init here * src/port.h: added rtrackUpdate * src/SDCCgen.c (emit_raw): try to track constant register contents * src/ds390/main.c, * src/hc08/main.c, * src/pic14/main.c, * src/pic16/main.c, * src/stm8/main.c, * src/z80/main.c (*_port): no register tracking * src/mcs51/rtrack.h: renamed rtrackUpdate to _mcs51_rtrackUpdate * src/mcs51/rtrack.c: enabled by default,   implemented suggested removals * src/mcs51/main.c: use _mcs51_rtrackUpdate * src/mcs51/gen.c (genSend, genPagedPointerGet): fixed bugs obscured by   peephole rules,   (genLeftShiftLiteral): disabled assert to fix bug 2224 * src/z80/z80.vcxproj, * src/z80/z80.vcxproj.filters: added peeph-tlcs90.def * support/regression/Makefile.in: clean host/spec.mk * support/regression/tests/bug-2222.c: enabled test for mcs51
git-svn-id: http://svn.code.sf.net/p/sdcc/code/trunk@8897 4a8a32a2-be11-0410-ad9d-d568d2c75423

+++ sdcc/sdas/asstm8/clean.mk
@@ -5,7 +5,7 @@ include $(top_builddir)/Makefile.common
 clean: mostlyclean
 	rm -f *.dep
 	rm -f $(ASOBJECTS) $(OBJDIR)/.stamp
-	-rmdir $(OBJDIR)
+	if [ -d $(OBJDIR) ]; then rmdir $(OBJDIR); fi
 
 # Deleting all files created by configuring or building the program
 # -----------------------------------------------------------------
